+++ lib/blizzard_api/wow/game_data/creature.rb
@@ -22,6 +22,7 @@ module BlizzardApi
       # Fetch all creature families
       #
       # @!macro request_options
+      # @option options [Boolean] :classic If set to true, this method will call the classic version
       #
       # @!macro response
       def families(options = {})
@@ -34,6 +35,7 @@ module BlizzardApi
       # @param id [Integer] Creature family id
       #
       # @!macro request_options
+      # @option options [Boolean] :classic If set to true, this method will call the classic version
       #
       # @!macro response
       def family(id, options = {})
@@ -46,6 +48,7 @@ module BlizzardApi
       # @param id [Integer] Creature family id
       #
       # @!macro request_options
+      # @option options [Boolean] :classic If set to true, this method will call the classic version
       #
       # @!macro response
       def family_media(id, options = {})
@@ -56,6 +59,7 @@ module BlizzardApi
       # Fetch all creature types
       #
       # @!macro request_options
+      # @option options [Boolean] :classic If set to true, this method will call the classic version
       #
       # @!macro response
       def types(options = {})
@@ -68,6 +72,7 @@ module BlizzardApi
       # @param id [Integer] Creature type id
       #
       # @!macro request_options
+      # @option options [Boolean] :classic If set to true, this method will call the classic version
       #
       # @!macro response
       def type(id, options = {})
@@ -80,6 +85,7 @@ module BlizzardApi
       # @param id [Integer] Creature type id
       #
       # @!macro request_options
+      # @option options [Boolean] :classic If set to true, this method will call the classic version
       #
       # @!macro response
       def display_media(id, options = {})
@@ -90,7 +96,7 @@ module BlizzardApi
 
       def endpoint_setup
         @endpoint = 'creature'
-        @namespace = endpoint_namespace :static
+        @namespace = :static
         @collection = 'achievements'
         @ttl = CACHE_TRIMESTER
       end