+++ lib/blizzard_api/wow/game_data/mythic_raid_leaderboard.rb
@@ -20,7 +20,7 @@ module BlizzardApi
       #
       # @!macro response
       def get(raid, faction, options = {})
-        opts = options.merge(namespace: endpoint_namespace(:dynamic), ttl: CACHE_DAY)
+        opts = options.merge(namespace: :dynamic, ttl: CACHE_DAY)
         api_request "#{base_url(:game_data)}/leaderboard/hall-of-fame/#{raid}/#{faction}", opts
       end
     end