+++ test/blizzard/wow/game_data/mythic_keystone_test.rb
@@ -27,7 +27,7 @@ module BlizzardApi
 
       def test_mythic_keystone_periods
         mythic_keystone_periods = @mythic_keystone_leaderboard.periods
-        assert_equal 66, mythic_keystone_periods[:periods].count
+        assert_equal 73, mythic_keystone_periods[:periods].count
       end
 
       def test_mythic_keystone_period