+++ test/blizzard/wow/game_data/race_test.rb
@@ -9,6 +9,9 @@ module BlizzardApi
         race = BlizzardApi::Wow::Race.new
         race_data = race.index
         assert_equal 23, race_data[:races].count
+
+        race_data = race.index
+        assert_equal 23, race_data[:races].count
       end
 
       def test_race_get